Week7/pyproject.toml
[tool.poetry]
name = "week6"
version = "0.1.0"
description = ""
authors = ["lucia <thisislucia1991@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
mysql-connector-python = "^8.4.0"
bcrypt = "3.1.7"
starlette = "^0.37.2"
itsdangerous = "^2.2.0"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
